move find_overlapping_roi_pairs to ophys_plane.py
so that it can be used in segmentation QC
